+++ CHANGELOG.md
@@ -23,6 +23,12 @@ different versioning scheme, following the Haskell community's
 * Fixed MSVC warning for deprecation of `std::result_of_t` in `/std:c++17`.
   ([Issue #1007](https://github.com/microsoft/bond/issues/1007))
 
+### C# ###
+
+* Fixed creation of duplicate default constructor when defining an empty struct
+  and using `--preview-constructor-parameters`.
+  ([Issue #963](https://github.com/microsoft/bond/issues/963))
+
 ### `gbc` and Bond compiler library ###
 * C++ codegen no longer supports MSVC 2013.